[hotfix][table-planner-blink] Merge tests of MiniBatchGroupWindowITCase into GroupWindowITCase
The Group Window Aggregate doesn't support mini-batch, so the test class doesn't make sense.
